fixes bug in the way wink web loads on the mac app.
the mac app now pings the now production site wink.blainemiller.xyz and
make sure that the site is live and then also offers people to host
their the site on their own server and link the mac app up to. come to
think of it the mac app actually can work as a micro browser up in the
menu bar now. enjoy ’(^,^)’